separating daemon from command-line tool
