+++ include/update_gps.h
@@ -1,0 +1,14 @@
+/******************************************************************************\
+** update_gps.h                                                               **
+** Interface file for the GPS update routine                                  **
+\******************************************************************************/
+
+#ifndef __UPDATE_GPS_H__
+#define __UPDATE_GPS_H__
+
+#include "libttwatch.h"
+
+/******************************************************************************/
+void do_update_gps(TTWATCH *watch);
+
+#endif  /* __UPDATE_GPS_H__ */
