+++ src/ttwatch.c
@@ -7,6 +7,10 @@
 #include "json.h"
 #include "log.h"
 #include "options.h"
+#include "misc.h"
+#include "firmware.h"
+#include "update_gps.h"
+#include "set_time.h"
 
 #include "libttwatch.h"
 
@@ -27,8 +31,6 @@
 #include <curl/curl.h>
 
 /*************************************************************************************************/
-
-#define TT_MANIFEST_ENTRY_UTC_OFFSET    (169)
 
 #define MANIFEST_TYPE_ENUM  (0)
 #define MANIFEST_TYPE_INT   (1)
@@ -97,21 +99,6 @@ struct
 };
 
 #define MANIFEST_DEFINITION_COUNT (sizeof(MANIFEST_DEFINITIONS) / sizeof(MANIFEST_DEFINITIONS[0]))
-
-/*************************************************************************************************/
-
-typedef struct
-{
-    uint8_t *data;
-    uint32_t length;
-} DOWNLOAD;
-
-typedef struct
-{
-    uint32_t id;
-    DOWNLOAD download;
-} FIRMWARE_FILE;
-
 
 /*************************************************************************************************/
 
@@ -207,182 +194,6 @@ void read_all_files(TTWATCH *watch)
 #endif  /* UNSAFE */
 
 /*****************************************************************************/
-static size_t write_download(char *ptr, size_t size, size_t nmemb, void *userdata)
-{
-    DOWNLOAD *download = (DOWNLOAD*)userdata;
-    /* reallocate storage for the data to fit the extra size */
-    size_t length = size * nmemb;
-    download->data = realloc(download->data, download->length + length);
-
-    /* store the new data and update the data length */
-    memcpy(download->data + download->length, ptr, length);
-    download->length += length;
-    return length;
-}
-
-/*****************************************************************************/
-static int download_file(const char *url, DOWNLOAD *download)
-{
-    int result;
-    CURL *curl = curl_easy_init();
-    if (!curl)
-    {
-        write_log(1, "Unable to initialise libcurl\n");
-        return 1;
-    }
-
-    download->data   = 0;
-    download->length = 0;
-
-    curl_easy_setopt(curl, CURLOPT_URL, url);
-    curl_easy_setopt(curl, CURLOPT_NOPROGRESS, 1);
-    curl_easy_setopt(curl, CURLOPT_USERAGENT, "curl/7.35.0");
-    curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 50);
-    curl_easy_setopt(curl, CURLOPT_TCP_KEEPALIVE, 1);
-    curl_easy_setopt(curl, CURLOPT_WRITEDATA, download);
-    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_download);
-
-    result = curl_easy_perform(curl);
-    curl_easy_cleanup(curl);
-    if (result != CURLE_OK)
-    {
-        write_log(1, "Unable to download file: %s (%d)\n", url, result);
-        return 1;
-    }
-    return 0;
-}
-
-/*****************************************************************************/
-char *replace(char *str, const char *old, const char *new)
-{
-    int old_len = strlen(old);
-    int new_len = strlen(new);
-    int correct = new_len - old_len;
-
-    char *ptr = strstr(str, old);
-    while (ptr)
-    {
-        const size_t offset = ptr - str;
-        const size_t new_str_len = (int)strlen(str) + correct + 1;
-        char *new_str = (char*)malloc(new_str_len);
-        memset(new_str, 0, new_str_len);
-
-        strncpy(new_str, str, offset);
-        strcat(new_str, new);
-        strcat(new_str, ptr + old_len);
-
-        free(str);
-        str = new_str;
-        ptr = strstr(str, old);
-    }
-
-    return str;
-}
-
-/*****************************************************************************/
-char *get_config_string(TTWATCH *watch, const char *name)
-{
-    char *start, *end;
-    char *url;
-    size_t length;
-    DOWNLOAD download = { 0, 0 };
-    json_value *value = 0;
-    unsigned i;
-
-    /* find the value of ConfigURL from the preferences file */
-    start = strstr(watch->preferences_file, "<ConfigURL>");
-    if (!start)
-        return NULL;
-
-    end = strstr(start, "</ConfigURL>");
-    if (!end)
-        return NULL;
-
-    /* find the start and end of the url value, skipping whitespace */
-    start += 11;
-    end -= 1;
-    while (isspace(*start))
-        ++start;
-    while (isspace(*end))
-        --end;
-
-    if (start >= end)
-        return NULL;
-
-    length = end - start + 1;
-    url = (char*)malloc(length + 1);
-    memcpy(url, start, length);
-    url[length] = 0;
-
-    /* download the config JSON file */
-    if (download_file(url, &download))
-    {
-        free(url);
-        return NULL;
-    }
-    free(url);
-    url = NULL;
-
-    /* attempt to parse the config JSON file */
-    value = json_parse((json_char*)download.data, download.length);
-    if (!value || (value->type != json_object))
-        goto cleanup;
-
-    /* loop through all the object entries to find the firmware entry */
-    for (i = 0; i < value->u.object.length; ++i)
-    {
-        if (strcasecmp(value->u.object.values[i].name, name) != 0)
-            continue;
-
-        if (value->u.object.values[i].value->type == json_string)
-            url = strdup(value->u.object.values[i].value->u.string.ptr);
-
-        break;
-    }
-
-cleanup:
-    json_value_free(value);
-    free(download.data);
-    return url;
-}
-
-
-/*****************************************************************************/
-void do_update_gps(TTWATCH *watch)
-{
-    DOWNLOAD download = {0};
-    char *url = 0;
-
-    url = get_config_string(watch, "service:ephemeris");
-    if (!url)
-    {
-        write_log(1, "Unable to get GPSQuickFix data URL\n");
-        return;
-    }
-    /* get 3 days ephemeris */
-    url = replace(url, "{DAYS}", "3");
-
-    /* download the data file */
-    write_log(0, "Downloading GPSQuickFix data file...\n");
-    if (download_file(url, &download))
-    {
-        free(download.data);
-        return;
-    }
-
-    write_log(0, "Writing file to watch...\n");
-    if (ttwatch_write_verify_whole_file(watch, TTWATCH_FILE_GPSQUICKFIX_DATA, download.data, download.length) == TTWATCH_NoError)
-    {
-        write_log(0, "GPSQuickFix data updated\n");
-        ttwatch_reset_gps_processor(watch);
-    }
-    else
-        write_log(1, "GPSQuickFix update failed\n");
-
-    free(download.data);
-}
-
-/*****************************************************************************/
 void do_get_time(TTWATCH *watch)
 {
     char timestr[64];
@@ -410,479 +221,6 @@ void do_get_time(TTWATCH *watch)
 }
 
 /*****************************************************************************/
-void do_set_time(TTWATCH *watch)
-{
-    struct tm tm_local;
-    time_t utc_time = time(NULL);
-
-    localtime_r(&utc_time, &tm_local);
-    int32_t offset = tm_local.tm_gmtoff;
-
-    time_t watch_utc_time;
-    if (ttwatch_get_watch_time(watch, &watch_utc_time) != TTWATCH_NoError)
-    {
-        write_log(1, "Unable to get watch time\n");
-        return;
-    }
-    /* if the UTC time of the watch is severely skewed, adjust the
-       offset so that the local time still shows up correctly, and
-       warn the user */
-    if (abs(watch_utc_time - utc_time) > 600)
-    {
-        write_log(1, "Watch has wrong UTC time. Use GPS to correct this.\n");
-        offset += utc_time - watch_utc_time;
-    }
-
-    if ((ttwatch_set_manifest_entry(watch, TT_MANIFEST_ENTRY_UTC_OFFSET, (uint32_t*)&offset) != TTWATCH_NoError) ||
-        (ttwatch_write_manifest(watch) != TTWATCH_NoError))
-    {
-        write_log(1, "Unable to set watch time\n");
-        return;
-    }
-
-    if (ttwatch_get_watch_time(watch, &utc_time) != TTWATCH_NoError)
-        write_log(1, "Unable to get watch time\n");
-}
-
-/*****************************************************************************/
-/* performs a 'mkdir -p', i.e. creates an entire directory tree */
-static void _mkdir(const char *dir)
-{
-    char tmp[256];
-    char *p = NULL;
-    size_t len;
-
-    snprintf(tmp, sizeof(tmp), "%s", dir);
-    len = strlen(tmp);
-    if(tmp[len - 1] == '/')
-        tmp[len - 1] = 0;
-    for(p = tmp + 1; *p; p++)
-    {
-        if(*p == '/')
-        {
-            *p = 0;
-            mkdir(tmp, 0755);
-            *p = '/';
-        }
-    }
-    mkdir(tmp, 0755);
-}
-
-/*****************************************************************************/
-typedef struct
-{
-    TTWATCH *watch;
-    OPTIONS *options;
-    uint32_t formats;
-
-} DGACallback;
-void do_get_activities_callback(uint32_t id, uint32_t length, void *cbdata)
-{
-    DGACallback *c = (DGACallback*)cbdata;
-    char filename[256] = {0};
-    uint8_t *data;
-    TTBIN_FILE *ttbin;
-    FILE *f;
-    struct tm timestamp;
-    uint32_t fmt1;
-    int i;
-
-    if (ttwatch_read_whole_file(c->watch, id, (void**)&data, 0) != TTWATCH_NoError)
-    {
-        write_log(1, "Unable to read activity file\n");
-        return;
-    }
-
-    /* parse the activity file */
-    ttbin = parse_ttbin_data(data, length);
-
-    if (ttbin)
-        gmtime_r(&ttbin->timestamp_local, &timestamp);
-    else
-    {
-        time_t t = time(NULL);
-        gmtime_r(&t, &timestamp);
-    }
-
-    /* create the directory name: [store]/[watch name]/[date] */
-    strcpy(filename, c->options->activity_store);
-    strcat(filename, "/");
-    if (ttwatch_get_watch_name(c->watch, filename + strlen(filename), sizeof(filename) - strlen(filename)) == TTWATCH_NoError)
-        strcat(filename, "/");
-    sprintf(filename + strlen(filename), "%04d-%02d-%02d",
-        timestamp.tm_year + 1900, timestamp.tm_mon + 1, timestamp.tm_mday);
-    _mkdir(filename);
-    chdir(filename);
-
-    /* create the file name */
-    if (ttbin)
-        sprintf(filename, "%s", create_filename(ttbin, "ttbin"));
-    else
-        sprintf(filename, "Unknown_%d-%d-%d_%d.ttbin", timestamp.tm_hour, timestamp.tm_min, timestamp.tm_sec, length);
-
-    /* write the ttbin file */
-    f = fopen(filename, "w+");
-    if (f)
-    {
-        uint8_t *data1;
-        fwrite(data, 1, length, f);
-
-        /* verify that the file was written correctly */
-        fseek(f, 0, SEEK_SET);
-        data1 = malloc(length);
-        if ((fread(data1, 1, length, f) != length) ||
-            (memcmp(data, data1, length) != 0))
-        {
-            write_log(1, "TTBIN file did not verify correctly\n");
-            if (ttbin)
-                free_ttbin(ttbin);
-            free(data);
-            free(data1);
-            return;
-        }
-        else
-        {
-            /* delete the file from the watch only if verification passed */
-            ttwatch_delete_file(c->watch, id);
-        }
-        free(data1);
-
-        fclose(f);
-    }
-    else
-        write_log(1, "Unable to write file: %s\n", filename);
-
-    if (!ttbin)
-    {
-        write_log(1, "Warning: Corrupt or unrecognisable activity file: %s\n", filename);
-        free(data);
-        return;
-    }
-
-    /* download the elevation data */
-    if (c->formats && ttbin->gps_records.count && !c->options->skip_elevation)
-    {
-        write_log(0, "Downloading elevation data\n");
-        download_elevation_data(ttbin);
-    }
-
-    /* export_formats returns the formats parameter with bits corresponding to failed exports cleared */
-    fmt1 = c->formats ^ export_formats(ttbin, c->formats);
-    if (fmt1)
-    {
-        write_log(1, "Unable to write file formats: ");
-        for (i = 0; i < OFFLINE_FORMAT_COUNT; ++i)
-        {
-            if (fmt1 & OFFLINE_FORMATS[i].mask)
-                write_log(1, "%s ", OFFLINE_FORMATS[i].name);
-        }
-        write_log(1, "\n");
-    }
-
-    /* don't run the post-processor as root */
-    if (c->options->post_processor && (getuid() != 0))
-    {
-        if (fork() == 0)
-        {
-            /* execute the post-processor */
-            execl(c->options->post_processor, c->options->post_processor, filename, (char*)0);
-        }
-    }
-
-    free_ttbin(ttbin);
-    free(data);
-}
-void do_get_activities(TTWATCH *watch, OPTIONS *options, uint32_t formats)
-{
-    DGACallback dgacallback = { watch, options, formats };
-    if (ttwatch_enumerate_files(watch, TTWATCH_FILE_TTBIN_DATA, do_get_activities_callback, &dgacallback) != TTWATCH_NoError)
-        write_log(1, "Unable to enumerate files\n");
-}
-
-/*****************************************************************************/
-/* find the firmware version from the downloaded firmware version XML file */
-uint32_t decode_latest_firmware_version(char *data)
-{
-    uint32_t latest_version = 0;
-    char *ptr = strstr(data, "<latestVersion>");
-    if (!ptr)
-        return 0;
-    ptr = strstr(ptr + 15, "<Major>");
-    if (!ptr)
-        return 0;
-    latest_version |= atoi(ptr + 7) << 16;
-    ptr = strstr(ptr, "<Minor>");
-    if (!ptr)
-        return 0;
-    latest_version |= atoi(ptr + 7) << 8;
-    ptr = strstr(ptr, "<Build>");
-    if (!ptr)
-        return 0;
-    latest_version |= atoi(ptr + 7);
-    return latest_version;
-}
-
-/*****************************************************************************/
-/* find a firmware URL from the downloaded firmware version XML file. Updates the
-   provided pointer so that this function can be called repeatedly until it fails,
-   to find the complete list of URLs */
-char *find_firmware_url(char **data)
-{
-    char *str;
-    char *ptr = strstr(*data, "<URL>");
-    if (!ptr)
-        return 0;
-    *data = ptr + 5;
-
-    ptr = strstr(*data, "</URL>");
-    if (!ptr)
-        return 0;
-
-    str = malloc(ptr - *data + 1);
-    memcpy(str, *data, ptr - *data);
-    str[ptr - *data] = 0;
-
-    *data = ptr + 6;
-
-    return str;
-}
-
-/*****************************************************************************/
-/* updates either one single file (if id is not zero), or all remaining files (if id is zero) */
-int update_firmware_file(TTWATCH *watch, FIRMWARE_FILE *files, int file_count, uint32_t id)
-{
-    int i;
-    for (i = 0; i < file_count; ++i)
-    {
-        if (id && (files[i].id != id))
-            continue;
-        else if (!id && ((files[i].id == TTWATCH_FILE_SYSTEM_FIRMWARE) || (files[i].id == TTWATCH_FILE_GPS_FIRMWARE) ||
-                         (files[i].id == TTWATCH_FILE_MANIFEST1)       || (files[i].id == TTWATCH_FILE_MANIFEST2) ||
-                         (files[i].id == TTWATCH_FILE_BLE_FIRMWARE)))
-        {
-            continue;
-        }
-
-        /* update the file; even if it fails, still update the next file */
-        write_log(0, "Updating firmware file: %08x ... ", files[i].id);
-        fflush(stdout);
-        if (ttwatch_write_verify_whole_file(watch, files[i].id, files[i].download.data, files[i].download.length) != TTWATCH_NoError)
-            write_log(0, "Failed\n");
-        else
-            write_log(0, "Done\n");
-    }
-
-    return 0;
-}
-
-/*****************************************************************************/
-void do_update_firmware(TTWATCH *watch, int force)
-{
-    uint32_t product_id;
-    char url[128];
-    DOWNLOAD download;
-    uint32_t current_version;
-    uint32_t current_ble_version;
-    uint32_t latest_version;
-    uint32_t latest_ble_version;
-    FIRMWARE_FILE *firmware_files = 0;
-    int file_count = 0;
-    int i;
-    char *ptr, *fw_url;
-    char *serial = 0;
-    char *fw_config_url = 0;
-    char *fw_base_url = 0;
-
-    product_id = watch->product_id;
-    current_version = watch->firmware_version;
-    current_ble_version = watch->ble_version;
-    serial = strdup(watch->serial_number);
-
-    /* find the firmware config URL */
-    fw_config_url = get_config_string(watch, "service:firmware");
-    if (!fw_config_url)
-    {
-        write_log(1, "Unable to determine firmware verion information\n");
-        return;
-    }
-    fw_config_url = replace(fw_config_url, "{PRODUCT_ID}", "%08X");
-
-    /* find the firmware base URL */
-    fw_base_url = strdup(fw_config_url);
-    ptr = strrchr(fw_base_url, '/');
-    if (!ptr || ((strlen(fw_base_url) - (ptr - fw_base_url)) < 3))
-    {
-        write_log(1, "Invalid firmware config URL\n");
-        goto cleanup;
-    }
-    *++ptr = '%';
-    *++ptr = 's';
-    *++ptr = 0;
-
-    /* download the firmware information file */
-    sprintf(url, fw_config_url, product_id);
-    if (download_file(url, &download))
-        goto cleanup;
-    /* null-terminate the data (yes, we lose the last byte of the file,
-       but since it's XML it's only the last closing bracket) */
-    download.data[download.length - 1] = 0;
-
-    /* get the latest firmware version */
-    latest_version = decode_latest_firmware_version((char*)download.data);
-    if (!latest_version)
-    {
-        write_log(1, "Unable to determine latest firmware version\n");
-        goto cleanup;
-    }
-
-    /* find the latest BLE version for the Multisport version*/
-    if (!IS_SPARK(watch->usb_product_id))
-    {
-        ptr = strstr((char*)download.data, "<BLE version=\"");
-        if (!ptr)
-        {
-            write_log(1, "Unable to determine latest BLE version\n");
-            goto cleanup;
-        }
-        latest_ble_version = strtoul(ptr + 14, NULL, 0);
-
-        /* check to see if we need to do anything */
-        if (!force && (latest_ble_version <= current_ble_version))
-            write_log(1, "Current BLE firmware is already at latest version\n");
-        else
-        {
-            write_log(0, "Current BLE Firmware Version: %u\n", current_ble_version);
-            write_log(0, "Latest BLE Firmware Version : %u\n", latest_ble_version);
-
-            /* find the download URL of the BLE firmware */
-            ptr = strstr(ptr, "URL=\"");
-            if (!ptr)
-            {
-                write_log(1, "Unable to determine BLE firmware download URL\n");
-                goto cleanup;
-            }
-            fw_url = ptr + 5;
-            ptr = strstr(fw_url, "\"");
-            if (!ptr)
-            {
-                write_log(1, "Unable to determine BLE firmware download URL\n");
-                goto cleanup;
-            }
-            *ptr = 0;
-
-            /* find the BLE firmware file ID */
-            ptr = strrchr(fw_url, '/');
-            if (!ptr)
-            {
-                write_log(1, "Unable to determine BLE file ID\n");
-                goto cleanup;
-            }
-
-            firmware_files = (FIRMWARE_FILE*)realloc(firmware_files, ++file_count * sizeof(FIRMWARE_FILE));
-
-            firmware_files[file_count - 1].id = strtoul(ptr + 1, NULL, 0);
-            firmware_files[file_count - 1].download.data   = 0;
-            firmware_files[file_count - 1].download.length = 0;
-
-            /* create the full URL and download the file */
-            sprintf(url, fw_base_url, product_id, fw_url);
-            write_log(0, "Download %s ... ", url);
-            if (download_file(url, &firmware_files[file_count - 1].download))
-            {
-                write_log(0, "Failed\n");
-                goto cleanup;
-            }
-            else
-                write_log(0, "Done\n");
-        }
-    }
-
-    /* check to see if we need to update the firmware */
-    if (!force && (latest_version <= current_version))
-        write_log(1, "Current firmware is already at latest version\n");
-    else
-    {
-        write_log(0, "Current Firmware Version: 0x%08x\n", current_version);
-        write_log(0, "Latest Firmware Version : 0x%08x\n", latest_version);
-
-        /* download all the firmware files from the server */
-        ptr = (char*)download.data;
-        while ((fw_url = find_firmware_url(&ptr)) != 0)
-        {
-            /* find the file ID within the file URL */
-            char *start = strrchr(fw_url, '/');
-            if (!start)
-            {
-                free(fw_url);
-                continue;
-            }
-
-            /* allocate space for the new file */
-            firmware_files = (FIRMWARE_FILE*)realloc(firmware_files, ++file_count * sizeof(FIRMWARE_FILE));
-
-            firmware_files[file_count - 1].id = strtoul(start + 1, NULL, 0);
-            firmware_files[file_count - 1].download.data   = 0;
-            firmware_files[file_count - 1].download.length = 0;
-
-            /* create the full URL and download the file */
-            sprintf(url, fw_base_url, product_id, fw_url);
-            free(fw_url);
-            write_log(0, "Download %s ... ", url);
-            if (download_file(url, &firmware_files[file_count - 1].download))
-            {
-                write_log(0, "Failed\n");
-                goto cleanup;
-            }
-            else
-                write_log(0, "Done\n");
-        }
-    }
-
-    if (file_count > 0)
-    {
-        /* update files in this order: 0x000000f0, 0x00010200, 0x00850000, 0x00850001, others */
-        update_firmware_file(watch, firmware_files, file_count, TTWATCH_FILE_BLE_FIRMWARE);
-        update_firmware_file(watch, firmware_files, file_count, TTWATCH_FILE_SYSTEM_FIRMWARE);
-        ttwatch_send_message_group_1(watch);   /* not sure why this is needed */
-        update_firmware_file(watch, firmware_files, file_count, TTWATCH_FILE_GPS_FIRMWARE);
-        ttwatch_send_message_group_1(watch);   /* not sure why this is needed */
-        update_firmware_file(watch, firmware_files, file_count, TTWATCH_FILE_MANIFEST1);
-        update_firmware_file(watch, firmware_files, file_count, TTWATCH_FILE_MANIFEST2);
-        update_firmware_file(watch, firmware_files, file_count, 0);
-
-        /* resetting the watch causes the watch to disconnect and
-           reconnect. This can take almost 90 seconds to complete */
-        ttwatch_reset_watch(watch);
-        write_log(0, "Waiting for watch to restart");
-        for (i = 0; i < 45; ++i)
-        {
-            sleep(2);
-            write_log(0, ".");
-        }
-        write_log(0, "\n");
-
-        write_log(0, "Firmware updated\n");
-    }
-
-cleanup:
-    /* free all the allocated data */
-    free(fw_config_url);
-    free(fw_base_url);
-    free(serial);
-    if (download.data)
-        free(download.data);
-    if (firmware_files)
-    {
-        while (file_count--)
-        {
-            if (firmware_files[file_count].download.data)
-                free(firmware_files[file_count].download.data);
-        }
-        free(firmware_files);
-    }
-}
-
-/*****************************************************************************/
 void do_get_watch_name(TTWATCH *watch)
 {
     char name[64];
@@ -900,25 +238,6 @@ void do_set_watch_name(TTWATCH *watch, const char *name)
     if (ttwatch_update_preferences_modified_time(watch) != TTWATCH_NoError)
         write_log(1, "Unable to write new watch name\n");
 }
-
-/*****************************************************************************/
-static void get_configured_formats_callback(const char *id, int auto_open, void *data)
-{
-    unsigned i;
-    for (i = 0; i < OFFLINE_FORMAT_COUNT; ++i)
-    {
-        if (strcasecmp(id, OFFLINE_FORMATS[i].name) == 0)
-            *(uint32_t*)data |= OFFLINE_FORMATS[i].mask;
-    }
-}
-static uint32_t get_configured_formats(TTWATCH *watch)
-{
-    uint32_t formats = 0;
-    if (ttwatch_enumerate_offline_formats(watch, get_configured_formats_callback, &formats) != TTWATCH_NoError)
-        write_log(1, "Unable to read configured formats\n");
-    return formats;
-}
-
 /*****************************************************************************/
 void do_list_formats(TTWATCH *watch)
 {
@@ -1898,162 +1217,6 @@ int list_devices_callback(TTWATCH *watch, void *data)
     return 1;
 }
 
-#ifdef DAEMON
-/*****************************************************************************/
-void daemon_watch_operations(TTWATCH *watch, OPTIONS *options)
-{
-    char name[32];
-    OPTIONS *new_options = copy_options(options);
-
-    /* make a copy of the options, and load any overriding
-       ones from the watch-specific conf file */
-    if (ttwatch_get_watch_name(watch, name, sizeof(name)) != TTWATCH_NoError)
-    {
-        char *filename = malloc(strlen(new_options->activity_store) + 1 + strlen(name) + 1 + 12 + 1);
-        sprintf(filename, "%s/%s/ttwatch.conf", new_options->activity_store, name);
-        load_conf_file(filename, new_options, LoadDaemonOperations);
-        free(filename);
-    }
-
-    /* perform the activities the user has requested */
-    if (new_options->get_activities)
-    {
-        uint32_t formats = get_configured_formats(watch);
-        if (!new_options->set_formats)
-            formats |= new_options->formats;
-        do_get_activities(watch, new_options, formats);
-    }
-
-    if (new_options->update_gps)
-        do_update_gps(watch);
-
-    if (new_options->update_firmware)
-        do_update_firmware(watch, 0);
-
-    if (new_options->set_time)
-        do_set_time(watch);
-
-    free_options(new_options);
-}
-
-/*****************************************************************************/
-int hotplug_attach_callback(struct libusb_context *ctx, struct libusb_device *dev,
-    libusb_hotplug_event event, void *user_data)
-{
-    OPTIONS *options = (OPTIONS*)user_data;
-    TTWATCH *watch = 0;
-
-    write_log(0, "Watch connected...\n");
-
-    if (ttwatch_open_device(dev, options->select_device ? options->device : 0, &watch) == TTWATCH_NoError)
-    {
-        daemon_watch_operations(watch, options);
-
-        write_log(0, "Finished watch operations\n");
-
-        ttwatch_close(watch);
-    }
-    else
-        write_log(0, "Watch not processed - does not match user selection\n");
-    return 0;
-}
-
-/*****************************************************************************/
-void daemonise(const char *user)
-{
-    pid_t pid, sid;
-    struct passwd *pwd;
-    struct group *grp;
-    FILE *file;
-    char *group;
-
-    /* we must be run as root */
-    if (getuid() != 0)
-    {
-        write_log(1, "This daemon must be run as root. The --runas=[USER[:GROUP]] parameter can be\n");
-        write_log(1, "used to run as an unprivileged user after daemon initialisation\n");
-        exit(1);
-    }
-
-    /* find the group in the user name */
-    group = 0;
-    if (user)
-        group = strchr(user, ':');
-    if (group)
-        *(group++) = 0;
-
-    /* perform the first fork */
-    pid = fork();
-    if (pid < 0)
-        exit(1);
-    if (pid > 0)
-        exit(0);
-
-    /* create a new session with no controlling terminal */
-    sid = setsid();
-    if (sid < 0)
-        _exit(1);
-
-    /* perform the second fork */
-    pid = fork();
-    if (pid)
-        _exit(0);
-
-    /* chdir to the root directory to prevent issues with removing inuse directories */
-    if (chdir("/") < 0)
-        _exit(1);
-
-    /* set umask to 0 since we don't know what it was before */
-    umask(0);
-
-    /* create the pid file */
-    file = fopen("/var/run/ttwatch.pid", "w");
-    if (file)
-    {
-        fprintf(file, "%d\n", getpid());
-        fclose(file);
-    }
-
-    /* retrieve information about the user we want to run as */
-    if (user)
-    {
-        pwd = getpwnam(user);
-        if (!pwd)
-        {
-            perror("getpwnam");
-            _exit(1);
-        }
-    }
-    if (group)
-    {
-        grp = getgrnam(group);
-        if (!grp)
-        {
-            perror("getgrnam");
-            _exit(1);
-        }
-        /* rewrite the ':' so when we do a ps, we see the original command line... */
-        *(group - 1) = ':';
-    }
-
-    /* make the log directory... */
-    mkdir("/var/log/ttwatch", 0755);
-    /* and make our user the owner */
-    if (user)
-        chown("/var/log/ttwatch", pwd->pw_uid, pwd->pw_gid);
-
-    /* drop privileges to that of the specified user */
-    if (user)
-    {
-        if (group)
-            setgid(grp->gr_gid);
-        else
-            setgid(pwd->pw_gid);
-        setuid(pwd->pw_uid);
-    }
-}
-#endif  /* DAEMON */
-
 /*****************************************************************************/
 void help(char *argv[])
 {
@@ -2075,9 +1238,6 @@ void help(char *argv[])
     write_log(0, "                               watch. Does NOT save the data before deleting it\n");
     write_log(0, "      --create-continuous-race [RACE] Create a continuously monitored race and\n");
     write_log(0, "                               uploads it to the watch (see below)\n");
-#ifdef DAEMON
-    write_log(0, "      --daemon               Run the program in daemon mode\n");
-#endif
 #ifdef UNSAFE
     write_log(0, "      --delete               Deletes a single file from the device\n");
 #endif
@@ -2103,9 +1263,6 @@ void help(char *argv[])
     write_log(0, "                               to/from the watch. Only used for debugging\n");
 #ifdef UNSAFE
     write_log(0, "  -r, --read=NUMBER          Reads a single file from the device\n");
-#endif
-#ifdef DAEMON
-    write_log(0, "      --runas=USER[:GROUP]   Run the daemon as the specified user, and\n");
 #endif
     write_log(0, "                               optionally as the specified group\n");
     write_log(0, "      --set-formats=LIST     Sets the list of file formats that are saved\n");
@@ -2208,18 +1365,6 @@ void help(char *argv[])
     write_log(0, "    specifies a race for running 400m in 1 minute, followed by 10km at 4min/km\n");
     write_log(0, "Note that the watch cannot differentiate between multiple segments, so it\n");
     write_log(0, "will not notify you when you have finished a segment.\n");
-#ifdef DAEMON
-    write_log(0, "\n");
-    write_log(0, "The program can be run as a daemon, which will automatically perform\n");
-    write_log(0, "the operations specified on the command line whenever a watch is\n");
-    write_log(0, "connected. The program must be run as root initially, but the --runas\n");
-    write_log(0, "parameter can be used to specify a user (and optionally a group) that the\n");
-    write_log(0, "program will drop privileges to after it has finished initialising.\n");
-    write_log(0, "Without this parameter specified, the program will continue to run as\n");
-    write_log(0, "root, which is not recommended for security reasons. Note that this\n");
-    write_log(0, "unprivileged user must have access to the USB devices, and write access\n");
-    write_log(0, "to the activity store location.\n");
-#endif
 }
 
 /*****************************************************************************/
@@ -2264,10 +1409,6 @@ int main(int argc, char *argv[])
         { "devices",        no_argument,       &options->list_devices,    1 },
         { "get-formats",    no_argument,       &options->list_formats,    1 },
         { "get-name",       no_argument,       &options->get_name,        1 },
-#ifdef DAEMON
-        { "daemon",         no_argument,       &options->daemon_mode,     1 },
-        { "runas",          required_argument, 0, 3   },
-#endif
         { "list-races",     no_argument,       &options->list_races,      1 },
         { "list-history",   no_argument,       &options->list_history,    1 },
         { "clear-data",     no_argument,       &options->clear_data,      1 },
@@ -2314,14 +1455,6 @@ int main(int argc, char *argv[])
             options->set_formats = 1;
             options->formats = parse_format_list(optarg);
             break;
-#ifdef DAEMON
-        case 3:     /* set daemon user */
-            options->run_as = 1;
-            if (options->run_as_user)
-                free(options->run_as_user);
-            options->run_as_user = strdup(optarg);
-            break;
-#endif
         case 4:     /* delete history entry */
             options->delete_history = 1;
             if (options->history_entry)
@@ -2444,89 +1577,6 @@ int main(int argc, char *argv[])
         if (options->activity_store)
             sprintf(options->activity_store, "%s/ttwatch", home);
     }
-
-#ifdef DAEMON
-    /* if daemon mode is requested ...*/
-    if (options->daemon_mode)
-    {
-        /* if the user has selected a device, make sure it's by name or serial number */
-        if (options->select_device)
-        {
-            if (isdigit(options->device[0]))
-            {
-                write_log(1, "Device selection in daemon mode must be by name or serial number\n");
-                free_options(options);
-                return 1;
-            }
-        }
-
-        /* reload the daemon operations from the config file. We don't load a per-user file */
-        load_conf_file("/etc/ttwatch", options, LoadDaemonOperations);
-
-        /* we have to include some useful functions, otherwise there's no point... */
-        if (!options->update_firmware && !options->update_gps && !options->get_activities && !options->set_time)
-        {
-            write_log(1, "To run as a daemon, you must include one or more of:\n");
-            write_log(1, "    --update-fw\n");
-            write_log(1, "    --update-gps\n");
-            write_log(1, "    --get-activities\n");
-            write_log(1, "    --set-time\n");
-            write_log(1, "    --auto (OR -a)\n");
-            free_options(options);
-            return 1;
-        }
-
-        /* become a daemon */
-        daemonise(options->run_as ? options->run_as_user : NULL);
-
-        /* we're a daemon, so open the log file and report that we have started */
-        set_log_location(LOG_VAR_LOG);
-        write_log(0, "Starting daemon.\n");
-
-        libusb_init(NULL);
-
-        /* setup hot-plug detection so we know when a watch is plugged in */
-        if (libusb_has_capability(LIBUSB_CAP_HAS_HOTPLUG))
-        {
-            int result;
-            if ((result = libusb_hotplug_register_callback(NULL, LIBUSB_HOTPLUG_EVENT_DEVICE_ARRIVED,
-                LIBUSB_HOTPLUG_ENUMERATE, TOMTOM_VENDOR_ID, TOMTOM_MULTISPORT_PRODUCT_ID,
-                LIBUSB_HOTPLUG_MATCH_ANY, hotplug_attach_callback, options, NULL)) != 0)
-            {
-                write_log(1, "Unable to register hotplug callback: %d\n", result);
-                _exit(1);
-            }
-
-            if ((result = libusb_hotplug_register_callback(NULL, LIBUSB_HOTPLUG_EVENT_DEVICE_ARRIVED,
-                LIBUSB_HOTPLUG_ENUMERATE, TOMTOM_VENDOR_ID, TOMTOM_SPARK_CARDIO_PRODUCT_ID,
-                LIBUSB_HOTPLUG_MATCH_ANY, hotplug_attach_callback, options, NULL)) != 0)
-            {
-                write_log(1, "Unable to register hotplug callback: %d\n", result);
-                _exit(1);
-            }
-
-            if ((result = libusb_hotplug_register_callback(NULL, LIBUSB_HOTPLUG_EVENT_DEVICE_ARRIVED,
-                LIBUSB_HOTPLUG_ENUMERATE, TOMTOM_VENDOR_ID, TOMTOM_SPARK_MUSIC_PRODUCT_ID,
-                LIBUSB_HOTPLUG_MATCH_ANY, hotplug_attach_callback, options, NULL)) != 0)
-            {
-                write_log(1, "Unable to register hotplug callback: %d\n", result);
-                _exit(1);
-            }
-
-            /* infinite loop - handle events every 10 seconds */
-            while (1)
-            {
-                libusb_handle_events_completed(NULL, NULL);
-                usleep(10000);
-            }
-        }
-        else
-            write_log(0, "System does not support hotplug notification\n");
-
-        free_options(options);
-        _exit(0);
-    }
-#endif  /* DAEMON */
 
     /* we need to do something, otherwise just show the help */
     if (
